requirements.txt
streamlit
numpy
pillow
matplotlib
pymongo
python-dotenv
dnspython